Split worker API and transfer handlers
